feat : update and checked working

+++ src/main/java/com/bank/repository/RoleRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface RoleRepo extends CrudRepository<Role, Long> {
+public interface RoleRepo extends CrudRepository<Role, Integer> {
 
     Optional<Role> findByName(String name);
 }
